pyproject.toml
[tool.poetry]
name = "wallabag-to-epub"
version = "0.1.0"
description = ""
authors = ["kahnwong <karn@karnwong.me>"]
readme = "README.md"
packages = [
    {include = "src"}
]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pypub3 = "^2.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
